ai/requirements.txt
python-dotenv==0.20.0
confluent-kafka==2.0.2
faust==1.10.4
numpy==1.23.2
Keras
tensorflow
scikit-image==0.20.0
matplotlib==3.7.1
python-dotenv==0.20.0
boto3==1.28.3
jsonschema
faust